requirements-bg-tool.txt
backgroundremover
Pillow
numpy
